pyproject.toml
[tool.poetry]
name = "secret_santa"
version = "0.1.0"
description = ""
authors = ["Dima Ulasovets <d.ulasovets@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.18.1"
django-dotenv = "^1.4.2"
psycopg2-binary = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
